+++ src/main/java/com/kakao/cafe/users/repository/UserRepository.java
@@ -1,4 +1,6 @@
-package com.kakao.cafe.users.domain;
+package com.kakao.cafe.users.repository;
+
+import com.kakao.cafe.users.domain.User;
 
 import java.util.List;
 import java.util.Optional;
