+++ src/test/java/com/kakao/cafe/users/repository/UserRepositoryTest.java
@@ -1,7 +1,8 @@
-package com.kakao.cafe.users.domain;
+package com.kakao.cafe.users.repository;
 
 import com.kakao.cafe.exception.repository.RequiredFieldNotFoundException;
 import com.kakao.cafe.exception.repository.UniqueFieldDuplicatedException;
+import com.kakao.cafe.users.domain.User;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
